+++ lib/src/element/sub/other/img_element.dart
@@ -1,13 +1,8 @@
+import 'dart:typed_data';
+
 import 'package:flutter/material.dart';
 
-import '../../../element_params/spwml_info.dart';
-import '../../../element_params/sub/other/img_element_params.dart';
-import '../../../element_params/super/spwml_params.dart';
-import '../../../enum/enum_img_type.dart';
-import '../../../enum/enum_spwml_params.dart';
-import '../../../enum/enum_spwml_element_type.dart';
-import '../../../style/spwml_font_style.dart';
-import '../../super/spwml_element.dart';
+import '../../../../simple_widget_markup.dart';
 
 ///
 /// Author Masahide Mori
@@ -56,21 +51,51 @@ class ImgElement extends SpWMLElement {
     elParams.p.repeat = params.containsKey(EnumSpWMLParams.repeat)
         ? params[EnumSpWMLParams.repeat]!
         : ImageRepeat.noRepeat;
+    elParams.p.clipParams.clipType =
+        params.containsKey(EnumSpWMLParams.clipType)
+            ? params[EnumSpWMLParams.clipType]!
+            : EnumClipType.none;
+    elParams.p.clipParams.borderRadius =
+        params.containsKey(EnumSpWMLParams.clipRadius)
+            ? params[EnumSpWMLParams.clipRadius]!
+            : BorderRadius.circular(8.0);
     return this;
+  }
+
+  /// Widget clipping if needed.
+  ///
+  /// * [w]: wrap target widget.
+  Widget _clipping(Widget w) {
+    if (elParams.p.clipParams.clipType == EnumClipType.none) {
+      return w;
+    } else if (elParams.p.clipParams.clipType == EnumClipType.oval) {
+      return ClipOval(
+          key: elParams.p.clipParams.key,
+          clipper: elParams.p.clipParams.clipper,
+          clipBehavior: elParams.p.clipParams.clipBehavior,
+          child: w);
+    } else {
+      return ClipRRect(
+          key: elParams.p.clipParams.key,
+          borderRadius: elParams.p.clipParams.borderRadius,
+          clipper: elParams.p.clipParams.clipperRRect,
+          clipBehavior: elParams.p.clipParams.clipBehavior,
+          child: w);
+    }
   }
 
   @override
   Widget getWidget(BuildContext context) {
     if (elParams.p.type == EnumImgType.asset) {
-      return Image.asset(
+      return _clipping(Image.asset(
         elParams.p.src ?? spwmlParams.p.text,
         key: elParams.p.key,
+        scale: elParams.p.scale,
         bundle: elParams.p.assetImgParams?.bundle,
         frameBuilder: elParams.p.frameBuilder,
         errorBuilder: elParams.p.errorBuilder,
         semanticLabel: elParams.p.semanticLabel,
         excludeFromSemantics: elParams.p.excludeFromSemantics,
-        scale: elParams.p.scale,
         width: elParams.p.width,
         height: elParams.p.height,
         color: elParams.p.color,
@@ -87,16 +112,17 @@ class ImgElement extends SpWMLElement {
         filterQuality: elParams.p.filterQuality,
         cacheWidth: elParams.p.cacheWidth,
         cacheHeight: elParams.p.cacheHeight,
-      );
-    } else {
-      return Image.network(
-        elParams.p.src ?? spwmlParams.p.text,
+      ));
+    } else if (elParams.p.type == EnumImgType.memory &&
+        elParams.p.bytes != null) {
+      return _clipping(Image.memory(
+        elParams.p.bytes!,
         key: elParams.p.key,
+        scale: elParams.p.scale,
         frameBuilder: elParams.p.frameBuilder,
         errorBuilder: elParams.p.errorBuilder,
         semanticLabel: elParams.p.semanticLabel,
         excludeFromSemantics: elParams.p.excludeFromSemantics,
-        scale: elParams.p.scale,
         width: elParams.p.width,
         height: elParams.p.height,
         color: elParams.p.color,
@@ -112,7 +138,40 @@ class ImgElement extends SpWMLElement {
         filterQuality: elParams.p.filterQuality,
         cacheWidth: elParams.p.cacheWidth,
         cacheHeight: elParams.p.cacheHeight,
-      );
+      ));
+    } else {
+      return _clipping(Image.network(
+        elParams.p.src ?? spwmlParams.p.text,
+        key: elParams.p.key,
+        scale: elParams.p.scale,
+        frameBuilder: elParams.p.frameBuilder,
+        errorBuilder: elParams.p.errorBuilder,
+        semanticLabel: elParams.p.semanticLabel,
+        excludeFromSemantics: elParams.p.excludeFromSemantics,
+        width: elParams.p.width,
+        height: elParams.p.height,
+        color: elParams.p.color,
+        opacity: elParams.p.opacity,
+        colorBlendMode: elParams.p.colorBlendMode,
+        fit: elParams.p.fit,
+        alignment: elParams.p.alignment,
+        repeat: elParams.p.repeat,
+        centerSlice: elParams.p.centerSlice,
+        matchTextDirection: elParams.p.matchTextDirection,
+        gaplessPlayback: elParams.p.gaplessPlayback,
+        isAntiAlias: elParams.p.isAntiAlias,
+        filterQuality: elParams.p.filterQuality,
+        cacheWidth: elParams.p.cacheWidth,
+        cacheHeight: elParams.p.cacheHeight,
+      ));
     }
+  }
+
+  /// (en)Set bytes image.
+  ///
+  /// (ja)バイトイメージを設定します。
+  /// * [image] The bytes image data.
+  void setBytesImage(Uint8List image) {
+    elParams.p.bytes = image;
   }
 }