+++ lib/src/element_params/sub/structure/span_params.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/gestures.dart';
-import 'package:flutter/material.dart';
+import 'package:flutter/rendering.dart';
 
 import '../text/text_params.dart';
 
@@ -28,6 +28,8 @@ class RichTextParams {
   StrutStyle? strutStyle;
   TextWidthBasis textWidthBasis = TextWidthBasis.parent;
   TextHeightBehavior? textHeightBehavior;
+  SelectionRegistrar? selectionRegistrar;
+  Color? selectionColor;
 }
 
 class TextSpanParams {