+++ lib/src/enum/enum_clip_type.dart
@@ -4,25 +4,25 @@ import '../spwml_exception.dart';
 ///
 /// Author Masahide Mori
 ///
-/// First edition creation date 2022-02-09 20:27:34
+/// First edition creation date 2023-02-05 20:40:02
 ///
-enum EnumImgType { network, asset, memory }
+enum EnumClipType { none, oval, rRect }
 
-extension EXTEnumImgType on EnumImgType {
+extension EXTEnumClipType on EnumClipType {
   /// Convert string name.
   String toStr() {
     return toString().split('.').last;
   }
 
   /// Throws [SpWMLException] : If the type is incorrect, Throws TypeException.
-  static EnumImgType fromStr(
+  static EnumClipType fromStr(
       String s, int lineStart, int lineEnd, SpWMLInfo? info) {
-    if (s == EnumImgType.network.toStr()) {
-      return EnumImgType.network;
-    } else if (s == EnumImgType.asset.toStr()) {
-      return EnumImgType.asset;
-    } else if (s == EnumImgType.memory.toStr()) {
-      return EnumImgType.memory;
+    if (s == EnumClipType.none.toStr()) {
+      return EnumClipType.none;
+    } else if (s == EnumClipType.oval.toStr()) {
+      return EnumClipType.oval;
+    } else if (s == EnumClipType.rRect.toStr()) {
+      return EnumClipType.rRect;
     } else {
       throw SpWMLException(
           EnumSpWMLExceptionType.typeException, lineStart, lineEnd, info);