+++ lib/src/enum/enum_indicator_type.dart
@@ -4,25 +4,23 @@ import '../spwml_exception.dart';
 ///
 /// Author Masahide Mori
 ///
-/// First edition creation date 2022-02-09 20:27:34
+/// First edition creation date 2023-02-06 19:55:21
 ///
-enum EnumImgType { network, asset, memory }
+enum EnumIndicatorType { circular, linear }
 
-extension EXTEnumImgType on EnumImgType {
+extension EXTEnumIndicatorType on EnumIndicatorType {
   /// Convert string name.
   String toStr() {
     return toString().split('.').last;
   }
 
   /// Throws [SpWMLException] : If the type is incorrect, Throws TypeException.
-  static EnumImgType fromStr(
+  static EnumIndicatorType fromStr(
       String s, int lineStart, int lineEnd, SpWMLInfo? info) {
-    if (s == EnumImgType.network.toStr()) {
-      return EnumImgType.network;
-    } else if (s == EnumImgType.asset.toStr()) {
-      return EnumImgType.asset;
-    } else if (s == EnumImgType.memory.toStr()) {
-      return EnumImgType.memory;
+    if (s == EnumIndicatorType.circular.toStr()) {
+      return EnumIndicatorType.circular;
+    } else if (s == EnumIndicatorType.linear.toStr()) {
+      return EnumIndicatorType.linear;
     } else {
       throw SpWMLException(
           EnumSpWMLExceptionType.typeException, lineStart, lineEnd, info);