+++ lib/src/enum/spwml_enum_all.dart
@@ -1,5 +1,7 @@
 export 'enum_btn_type.dart';
 export 'enum_img_type.dart';
+export 'enum_clip_type.dart';
+export 'enum_indicator_type.dart';
 export 'enum_spwml_element_type.dart';
 export 'enum_spwml_params.dart';
 export 'enum_textfield_params.dart';